More additions
Changed how strings are set in pre2, now with '@from "<**args>"' super and '@with nullsender' super. Additions to config file for mundane stateents that aren't quite errors but shouldn't be happening, alerts toggleable in code and from conf file, still need to add ability to change config throughout code using '@load config <filename>' 

Note: If there is a stew alert that doesn't change the console prefix based on a conf file, don't change any code, that is meant to be there, it's for things concerning the stew compiler not the 'boiled stew' aka compiled code!